fixing deriv_operators:
the stencils were wrong in the previous commit for the non-periodic_FD
operators. It is fixed and non-periodic FD operators are incorporated
in the unit tests.

NOTE: that in the 1d and 2d case of these tests that are checked against
an analytic solution, the test is passed when we check if the maximim abs
error is smaller than the local truncation error. However, if the test is
done exactly as the periodic case (compare with the analytic sol at each pt)
it fails.

Signed-off-by: Thanasis Giannakopoulos <giannakopoulos.th@gmail.com>